requirements.txt
# Renewable Baseload Analysis - Dependencies
# Install with: pip install -r requirements.txt

# Core data analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Statistical modeling
statsmodels>=0.14.0

# Network analysis
networkx>=3.1

# Mapping (required for Figure 3)
cartopy>=0.21.0

# Wavelet analysis
PyWavelets>=1.4.0

# ENTSO-E API (for data download only)
entsoe-py>=0.6.0

# Elexon API (for UK data download only)
elexonpy>=1.0.0

# ERA5 reanalysis (optional, for validation only)
# xarray>=2023.1.0
# cdsapi>=0.6.0
# netCDF4>=1.6.0
